refactor: extract useSidebar composable — remove ~20 lines of duplicated sidebar logic from BrowseView and GraphView
